file reorganization & and setup local environment
