--- message/src/messagesQueue/index.js
@@ -1,4 +1,0 @@
-const dispatcherMessage = require("./dispatcherMessage.js");
-const enqueueMessage = require("./enqueueMessage.js");
-
-module.exports = { dispatcherMessage, enqueueMessage };
